Login searches for account in LocalStorage
the login function now searches for an account in LocalStorage. If the account exists and the username and password match, the alert will say that the user is in the database. Note: I don't think this is similar to pulling data from a database. This is just so we can see how to get data from the input boxes