currently available Inserts for MySQL Tables
Some of the inserts only have 50 or 25 tuples. the ones with the most are the inserts for the activity table and the tracksactivities tables. the tracksactivities table has close to 1000 tuples in it.
The inserts are made with the assumption that the example inserts from the create database SQL file are already in the database so a lot of the inserts using userIDs start from 2 instead of 1.